v0.3.1 - Airplane Rotation Redo
Scrapped the previous airplane rotation code and working on an improved solution. Implemented a new coordinate system for mousePosition tracking on the game screen-- set the origin at the center of the screen, instead of at the bottom left, so we can use four quadrants for plane rotation and calculate distance from the mouse to the airplane.