v0.9.1 - Right Turn Crank Operable
Turn cranks work very well via mouse. Can't get the code to determine whether the crank is turning clockwise or counter-clockwise via player control, but for the demo we're fine without it. Added random rotation goal so the crank can be turned a few times before it no longer needs to be turned.

Still needs to be hooked up to player flight and part durability, and then copied over to the other wing/crank.