v0.10.0 - Smash Engine Script Added
The engine can be smashed to be repaired, with some screenshake included and the gauge partially hooked up. The durability system is in the middle of being heavily refactored: (1) Rather than all durability components being managed by the airplane, each component part is responsible for handing it's own durability; (2) the component script sends only 1 flag to the airplane script-- whether the part works or not, so everything else is handled internally; (3) using that 1 flag, the airplane determines how the airplane should handle mid-flight and what sort of juicy UI effects should occur.